Make it possible to start all services in sanbase
