+++ lib/sanbase/scrapers_supervisor.ex
@@ -1,17 +1,7 @@
 defmodule Sanbase.Application.ScrapersSupervisor do
-  use Application
-
-  import Supervisor.Spec
   import Sanbase.ApplicationUtils
 
-  # See https://hexdocs.pm/elixir/Application.html
-  # for more information on OTP Applications
-  def start(_type, _args) do
-    if Code.ensure_loaded?(Envy) do
-      Envy.auto_load()
-    end
-
-    # Define workers and child supervisors to be supervised
+  def children() do
     children = [
       # Start the Task Supervisor
       {Task.Supervisor, [name: Sanbase.TaskSupervisor]},
@@ -105,16 +95,6 @@ defmodule Sanbase.Application.ScrapersSupervisor do
       max_seconds: 1
     ]
 
-    # Add error tracking through sentry
-    :ok = :error_logger.add_report_handler(Sentry.Logger)
-
-    Supervisor.start_link(children, opts)
-  end
-
-  # Tell Phoenix to update the endpoint configuration
-  # whenever the application is updated.
-  def config_change(changed, _new, removed) do
-    SanbaseWeb.Endpoint.config_change(changed, removed)
-    :ok
+    {children, opts}
   end
 end
