+++ lib/sanbase/web_supervisor.ex
@@ -1,15 +1,7 @@
 defmodule Sanbase.Application.WebSupervisor do
-  use Application
-
   import Sanbase.ApplicationUtils
 
-  # See https://hexdocs.pm/elixir/Application.html
-  # for more information on OTP Applications
-  def start(_type, _args) do
-    if Code.ensure_loaded?(Envy) do
-      Envy.auto_load()
-    end
-
+  def children() do
     # Define workers and child supervisors to be supervised
     children = [
       # Start the Postgres Ecto repository
@@ -52,20 +44,8 @@ defmodule Sanbase.Application.WebSupervisor do
 
     children = children |> normalize_children()
 
-    # See https://hexdocs.pm/elixir/Supervisor.html
-    # for other strategies and supported options
     opts = [strategy: :one_for_one, name: Sanbase.WebSupervisor, max_restarts: 5, max_seconds: 1]
 
-    # Add error tracking through sentry
-    :ok = :error_logger.add_report_handler(Sentry.Logger)
-
-    Supervisor.start_link(children, opts)
-  end
-
-  # Tell Phoenix to update the endpoint configuration
-  # whenever the application is updated.
-  def config_change(changed, _new, removed) do
-    SanbaseWeb.Endpoint.config_change(changed, removed)
-    :ok
+    {children, opts}
   end
 end
